Allow fallback-configuration; improve insert_final_newline-implementation
The introduction of the 'auto'-setting is a preparation for future
updates, it also makes the whole implementation less intrusive and more
consistent, since some configuration values now used during the whole
lifetime of the TextBuffer.

The insert_final_newline-implementation still suffers from intrusive
actions of the whitespace-plugin. A workaround could be to disable the
entire whitespace-plugin in favor of atom-editorconfig.